requirements.txt
absl-py==2.1.0
asgiref==3.8.1
contourpy==1.2.1
cycler==0.12.1
Django==5.1
fonttools==4.53.1
h5py==3.11.0
image==1.5.33
joblib==1.4.2
keras==3.4.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.9.1.post1
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
